+++ demo/meeting/src/components/Table/TableRow.tsx
@@ -1,0 +1,10 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import React, { FC } from 'react';
+
+export const TableRow: FC<{}> = ({ children }) => {
+  return <tr>{children}</tr>;
+}
+
+export default TableRow;