+++ demo/meeting/src/views/AppCheckerDetail/Styled.tsx
@@ -1,0 +1,20 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import styled from 'styled-components';
+
+export const StyledLayout = styled.main`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  max-width: 85rem;
+  padding: 2rem;
+  margin: auto;
+
+  @media (max-width: 760px) {
+    border-right: unset;
+    align-items: unset;
+    justify-content: unset;
+  }
+`;