+++ src/components/ui/Badge/Styled.tsx
@@ -14,9 +14,17 @@ export const StyledBadge = styled.span<Partial<BadgeProps>>`
   color: ${props => props.theme.colors.greys.white};
   font-size: 0.65rem;
   background-color: ${props =>
-    props.status === 'alert'
-      ? props.theme.colors.error.primary
-      : props.theme.colors.greys.grey60};
+    props.status === 'alert' ?
+    props.theme.colors.error.primary
+    :
+    props.status === 'default' ?
+      props.theme.colors.greys.grey60
+      :
+      props.status === 'success' ?
+        props.theme.colors.success.primary
+        :
+        props.theme.colors.warning.primary    
+  };
 
   ${baseSpacing};
   ${baseStyles}