+++ tst/components/ui/Badge/index.test.tsx
@@ -15,6 +15,7 @@ describe('Badge', () => {
         const element = getByTestId('badge');
         expect(element).toBeInTheDocument();
         expect(element).toHaveTextContent("100");
+        expect(element).toHaveStyle(`background-color: #616672`);
     });
 
     it('should render a Badge with a string value and default status', () => {
@@ -23,6 +24,7 @@ describe('Badge', () => {
         const element = getByTestId('badge');
         expect(element).toBeInTheDocument();
         expect(element).toHaveTextContent("100+");
+        expect(element).toHaveStyle(`background-color: #616672`);
     });
 
     it('should render a Badge with a number value and alert status', () => {
@@ -31,5 +33,23 @@ describe('Badge', () => {
         const element = getByTestId('badge');
         expect(element).toBeInTheDocument();
         expect(element).toHaveTextContent("50");
+        expect(element).toHaveStyle(`background-color: #C52000`);
+    });
+
+    it('should render a Badge with a number value and success status', () => {
+        const component = <Badge value={50} status={'success'} />
+        const { getByTestId } = renderWithTheme(lightTheme, component);
+        const element = getByTestId('badge');
+        expect(element).toBeInTheDocument();
+        expect(element).toHaveStyle(`background-color: #067000`);
+    });
+
+    it('should render a Badge with a number value and warning status', () => {
+        const component = <Badge value={50} status={'warning'} />
+        const { getByTestId } = renderWithTheme(lightTheme, component);
+        const element = getByTestId('badge');
+        expect(element).toBeInTheDocument();
+        expect(element).toHaveTextContent("50");
+        expect(element).toHaveStyle(`background-color: #F9CC09`);
     });
 });