+++ tst/snapshots/Badge/index.test.tsx
@@ -13,4 +13,16 @@ describe('Badge', () => {
       const image = await page.screenshot();
       expect(image).toMatchImageSnapshot();
     });
+  
+    it('success status', async () => {
+      await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge&knob-status=success');
+      const image = await page.screenshot();
+      expect(image).toMatchImageSnapshot();
+    });
+  
+    it('warning status', async () => {
+      await page.goto('http://host.docker.internal:9009/iframe.html?id=ui-components-badge--basic-badge&knob-status=warning');
+      const image = await page.screenshot();
+      expect(image).toMatchImageSnapshot();
+    });
 });