+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `useLocalAudioInputActivityPreview` hook for direct access to microphone input value
+- Add app readiness check
 
 ### Changed